+++ src/View/Chart.java
@@ -72,56 +72,63 @@ public class Chart extends ApplicationFrame{
 		     }
 		   }
 	       if (metric.equals("total impression cost")) {
-	         ArrayList<Double> ticList = calculator.getTotalImpressionCostList();
-			 int count = 1;
+	         ArrayList<Float> ticList = calculator.getTotalImpressionCostList();
+			     int count = 1;
+           
 			 for (double i : ticList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
 			 }	
 		   }
 	       if (metric.equals("total click cost")) {
-	         ArrayList<Double> tccList = calculator.getTotalClickCostList();
-		     int count = 1;
+	         ArrayList<Float> tccList = calculator.getTotalClickCostList();
+		       int count = 1;
+           
 			 for (double i : tccList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
 		     }	
 		   }
 	       if (metric.equals("ctr")) {
-		     ArrayList<Double> ctrList = calculator.getCtrList();
-		     int count = 1;
+		       ArrayList<Float> ctrList = calculator.getCtrList();
+		       int count = 1;
+
 			 for (double i : ctrList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
 		     }
 		   }
 	       if (metric.equals("cpa")) {
-		     ArrayList<Double> cpaList = calculator.getCpaList();
-		     int count = 1;
+		       ArrayList<Float> cpaList = calculator.getCpaList();
+		       int count = 1;
+           
 			 for (double i : cpaList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
 		     }
 		   }
 	       if (metric.equals("cpc")) {
-		     ArrayList<Double> cpcList = calculator.getCpcList();
-		     int count = 1;
+		       ArrayList<Float> cpcList = calculator.getCpcList();
+		       int count = 1;
+           
 			 for (double i : cpcList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
 		     }
 		   }
 	       if (metric.equals("cpm")) {
-		     ArrayList<Double> cpmList = calculator.getCpmList();
-		     int count = 1;
+		       ArrayList<Float> cpmList = calculator.getCpmList();
+		       int count = 1;
+           
 			 for (double i : cpmList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
 		     }
 		   }
 	       if (metric.equals("bounce rate")) {
-		     ArrayList<Double> brList = calculator.getBrList();
-		     int count = 1;
+		       ArrayList<Float> brList = calculator.getBrList();
+		       int count = 1;
+           
 			 for (double i : brList) {
 			   dataset.addValue(i, metric, String.valueOf(count));
 			   count = count + 1;
